# Implement rules using a trie data structure.