--- a/lib/clean-host.js
+++ b/lib/clean-host.js
@@ -16,13 +16,14 @@ var URL = require('url');
 // scheme      = ALPHA *( ALPHA / DIGIT / "+" / "-" / "." )
 var hasPrefixRE = /^(([a-z][a-z0-9+.-]*)?:)?\/\//;
 var invalidHostnameChars = /[^A-Za-z0-9.-]/;
+var trailingDotsRE = /[.]+$/g;
 
 //@see https://github.com/oncletom/tld.js/issues/95
 function rtrim(value) {
-  return String(value).replace(/[.]+$/g, '');
+  return String(value).replace(trailingDotsRE, '');
 }
 
-module.exports = function cleanHostValue(value){
+module.exports = function cleanHostValue(value) {
   value = String(value).trim().toLowerCase();
 
   var parts = URL.parse(hasPrefixRE.test(value) ? value : '//' + value, null, true);