--- a/lib/clean-host.js
+++ b/lib/clean-host.js
@@ -23,7 +23,11 @@ function rtrim(value) {
   return String(value).replace(trailingDotsRE, '');
 }
 
-module.exports = function cleanHostValue(value) {
+module.exports = function cleanHostValue(value, isHostClean) {
+  if (isHostClean) {
+    return value;
+  }
+
   value = String(value).trim().toLowerCase();
 
   var parts = URL.parse(hasPrefixRE.test(value) ? value : '//' + value, null, true);