--- a/lib/clean-host.js
+++ b/lib/clean-host.js
@@ -18,21 +18,21 @@ var hasPrefixRE = /^(([a-z][a-z0-9+.-]*)?:)?\/\//;
 var invalidHostnameChars = /[^A-Za-z0-9.-]/;
 var trailingDotsRE = /[.]+$/g;
 
-//@see https://github.com/oncletom/tld.js/issues/95
+// @see https://github.com/oncletom/tld.js/issues/95
 function rtrim(value) {
   return String(value).replace(trailingDotsRE, '');
 }
 
-module.exports = function cleanHostValue(value, isHostClean) {
-  if (isHostClean) {
-    return value;
-  }
-
+module.exports = function cleanHostValue(value) {
   value = String(value).trim().toLowerCase();
 
   var parts = URL.parse(hasPrefixRE.test(value) ? value : '//' + value, null, true);
 
-  if (parts.hostname && !invalidHostnameChars.test(parts.hostname)) { return rtrim(parts.hostname); }
-  if (!invalidHostnameChars.test(value)) { return rtrim(value); }
+  if (parts.hostname && !invalidHostnameChars.test(parts.hostname)) {
+    return rtrim(parts.hostname);
+  } else if (!invalidHostnameChars.test(value)) {
+    return rtrim(value);
+  }
+
   return '';
 };