--- a/lib/domain.js
+++ b/lib/domain.js
@@ -1,9 +1,6 @@
-var Rule = require('./rule.js');
 var isValid = require('./is-valid.js');
 var cleanHostValue = require('./clean-host.js');
-var extractTldFromHost = require('./from-host.js');
-var getCandidateRule = require('./canditate-rule.js');
-var getRulesForTld = require('./tld-rules.js');
+var getPublicSuffix = require('./public-suffix.js');
 
 /**
  * Detects the domain based on rules and upon and a host string
@@ -12,26 +9,43 @@ var getRulesForTld = require('./tld-rules.js');
  * @param {string} host
  * @return {String}
  */
-module.exports = function getDomain (allRules, validHosts, host) {
-  var domain = null, hostTld, rules, rule;
+module.exports = function getDomain(allRules, validHosts, host, dirtyHost = true) {
   var _validHosts = validHosts || [];
+  var cleanHost = dirtyHost ? cleanHostValue(host) : host;
 
-  if (isValid(_validHosts, host) === false) {
+  if (isValid(_validHosts, cleanHost) === false) {
     return null;
   }
 
-  host = cleanHostValue(host);
-  hostTld = extractTldFromHost(host);
-  rules = getRulesForTld(allRules, hostTld, new Rule({"firstLevel": hostTld, "isHost": _validHosts.indexOf(hostTld) !== -1}));
-  rule = getCandidateRule(host, rules);
+  // Check if `host` ends with '.' followed by one host specified in validHosts.
+  for (var i = 0; i < validHosts.length; i++) {
+    var vhost = validHosts[i];
+    if (cleanHost.endsWith(vhost) && (
+          cleanHost.length === vhost.length ||
+          cleanHost[cleanHost.length - vhost.length - 1] === '.')) {
+      return vhost;
+    }
+  }
 
-  if (rule === null) {
+  var suffix = getPublicSuffix(allRules, cleanHost, false);
+  if (suffix === null) {
+    // TODO - shouldn't it be null?
+    // Otherwise 'should return the known valid host' fails
+    // return cleanHost;
     return null;
   }
 
-  host.replace(new RegExp(rule.getPattern()), function (m, d) {
-    domain = d;
-  });
+  if (suffix.length === cleanHost.length) {
+    return null;
+  }
 
-  return domain;
+  // google.fr (length 9)
+  // suffix = fr (length 2)
+  // 5 = 9 - 2 - 1 (ignore the dot) - 1 (zero-based indexing)
+  var lastDotBeforeSuffixIndex = cleanHost.lastIndexOf('.', cleanHost.length - suffix.length - 2);
+  if (lastDotBeforeSuffixIndex === -1) {
+    return cleanHost;
+  }
+
+  return cleanHost.substring(lastDotBeforeSuffixIndex + 1);
 };