--- a/lib/domain.js
+++ b/lib/domain.js
@@ -9,9 +9,9 @@ var getPublicSuffix = require('./public-suffix.js');
  * @param {string} host
  * @return {String}
  */
-module.exports = function getDomain(allRules, validHosts, host, dirtyHost = true) {
+module.exports = function getDomain(allRules, validHosts, host, dirtyHost) {
   var _validHosts = validHosts || [];
-  var cleanHost = dirtyHost ? cleanHostValue(host) : host;
+  var cleanHost = (dirtyHost || true) ? cleanHostValue(host) : host;
 
   if (isValid(_validHosts, cleanHost) === false) {
     return null;