--- a/lib/domain.js
+++ b/lib/domain.js
@@ -20,7 +20,7 @@ module.exports = function getDomain(allRules, validHosts, host, dirtyHost) {
   // Check if `host` ends with '.' followed by one host specified in validHosts.
   for (var i = 0; i < validHosts.length; i++) {
     var vhost = validHosts[i];
-    if (cleanHost.endsWith(vhost) && (
+    if (cleanHost.indexOf(vhost) === (cleanHost.length - vhost.length) && (
           cleanHost.length === vhost.length ||
           cleanHost[cleanHost.length - vhost.length - 1] === '.')) {
       return vhost;