--- a/lib/domain.js
+++ b/lib/domain.js
@@ -9,9 +9,9 @@ var getPublicSuffix = require('./public-suffix.js');
  * @param {string} host
  * @return {String}
  */
-module.exports = function getDomain(allRules, validHosts, host, dirtyHost) {
+module.exports = function getDomain(allRules, validHosts, host, isHostClean) {
   var _validHosts = validHosts || [];
-  var cleanHost = (dirtyHost || true) ? cleanHostValue(host) : host;
+  var cleanHost = cleanHostValue(host, isHostClean);
 
   if (isValid(_validHosts, cleanHost) === false) {
     return null;
@@ -27,7 +27,7 @@ module.exports = function getDomain(allRules, validHosts, host, dirtyHost) {
     }
   }
 
-  var suffix = getPublicSuffix(allRules, cleanHost, false);
+  var suffix = getPublicSuffix(allRules, cleanHost, true);
   if (suffix === null) {
     // TODO - shouldn't it be null?
     // Otherwise 'should return the known valid host' fails