--- a/lib/from-host.js
+++ b/lib/from-host.js
@@ -4,6 +4,11 @@
  * @param {string} host
  * @return {String}
  */
-module.exports = function extractTldFromHost(host){
-  return host.split('.').pop();
+module.exports = function extractTldFromHost(host) {
+  var lastDotIndex = host.lastIndexOf('.');
+  if (lastDotIndex === -1) {
+    return null;
+  }
+
+  return host.substring(lastDotIndex + 1);
 };