--- a/lib/from-host.js
+++ b/lib/from-host.js
@@ -1,3 +1,5 @@
+"use strict";
+
 /**
  * Utility to extract the TLD from a host string
  *
@@ -10,5 +12,5 @@ module.exports = function extractTldFromHost(host) {
     return null;
   }
 
-  return host.substring(lastDotIndex + 1);
+  return host.substr(lastDotIndex + 1);
 };