--- a/lib/is-valid.js
+++ b/lib/is-valid.js
@@ -1,3 +1,5 @@
+"use strict";
+
 /**
  * Checking if a host string is valid
  * It's usually a preliminary check before trying to use getDomain or anything else
@@ -8,6 +10,6 @@
  * @param host {String}
  * @return {Boolean}
  */
-module.exports = function isValid (validHosts, host) {
+module.exports = function isValid(validHosts, host) {
   return typeof host === 'string' && (validHosts.indexOf(host) !== -1 || (host.indexOf('.') !== -1 && host[0] !== '.'));
 };