--- a/lib/parsers/publicsuffix-org.js
+++ b/lib/parsers/publicsuffix-org.js
@@ -1,61 +1,10 @@
 "use strict";
 
+var punycode = require('punycode');
+var SuffixTrie = require('../suffix-trie.js');
+
 var PublicSuffixOrgParser = {};
 
-var punycode = require('punycode');
-
-/**
- * Parse a one-domain-per-line file
- *
- * @param body {String}
- * @return {Array}
- */
-PublicSuffixOrgParser.parse = function (body){
-  return (body+'')
-    .split(/\n/m)
-    .filter(PublicSuffixOrgParser.filterRow)
-    .map(PublicSuffixOrgParser.domainBuilder);
-};
-
-/**
- * Returns a rule based on string analysis
- *
- * @param rule {PublicSuffixRule}
- */
-PublicSuffixOrgParser.domainBuilder = function (row){
-  var rule = {};
-  
-  row = punycode.toASCII(row.trim());
-
-  //setting initial rule
-  rule.source = row;
-
-  //exception
-  row = row.replace(/^!(.+)$/, function(m, tld){
-    rule.exception = true;
-
-    return tld;
-  });
-
-  //wildcard
-  row = row.replace(/^(\*\.)(.+)$/, function(m, dummy, tld){
-    rule.wildcard = true;
-
-    return tld;
-  });
-
-  //splitting domains
-  row.replace(/^((.+)\.)?([^\.]+)$/, function(m, dummy, secondLevel, firstLevel){
-    rule.firstLevel = firstLevel;
-
-    if (secondLevel){
-      rule.secondLevel = secondLevel;
-    }
-
-  });
-
-  return rule;
-};
 
 /**
  * Filters a commented or empty line
@@ -63,8 +12,62 @@ PublicSuffixOrgParser.domainBuilder = function (row){
  * @param row {String}
  * @return {String|null}
  */
-PublicSuffixOrgParser.filterRow = function (row) {
-  return (/^\/\//).test(row) ? null : row;
+function keepOnlyRules(row) {
+  var trimmed = row.trim();
+  if (!trimmed || trimmed.startsWith('//')) {
+    return null;
+  }
+
+  // TODO - Ignore leading or trailing dot
+
+  return trimmed;
+}
+
+
+/**
+ * Returns a rule based on string analysis
+ *
+ * @param rule {PublicSuffixRule}
+ */
+function domainBuilder(row) {
+  var rule = {
+    exception: false,
+    source: null,
+    parts: null,
+  };
+
+  var spaceIndex = row.indexOf(' ');
+  // Only read line up to the first white-space
+  row = punycode.toASCII(row);
+
+  //setting initial rule
+  rule.source = row;
+
+  // exceptions
+  if (row.startsWith('!')) {
+    row = row.substring(1);
+    rule.exception = true;
+  }
+
+  rule.parts = row.split('.').reverse();
+
+  return rule;
+}
+
+
+/**
+ * Parse a one-domain-per-line file
+ *
+ * @param body {String}
+ * @return {Array}
+ */
+PublicSuffixOrgParser.parse = function (body) {
+  return new SuffixTrie((body + '')
+    .split('\n')
+    .map(keepOnlyRules)
+    .filter(function (r) { return r !== null; })
+    .map(domainBuilder));
 };
+
 
 module.exports = PublicSuffixOrgParser;