--- a/lib/parsers/publicsuffix-org.js
+++ b/lib/parsers/publicsuffix-org.js
@@ -14,7 +14,7 @@ var PublicSuffixOrgParser = {};
  */
 function keepOnlyRules(row) {
   var trimmed = row.trim();
-  if (!trimmed || trimmed.startsWith('//')) {
+  if (!trimmed || trimmed.indexOf('//') === 0) {
     return null;
   }
 
@@ -44,7 +44,7 @@ function domainBuilder(row) {
   rule.source = row;
 
   // exceptions
-  if (row.startsWith('!')) {
+  if (row[0] === '!') {
     row = row.substring(1);
     rule.exception = true;
   }