--- a/lib/parsers/publicsuffix-org.js
+++ b/lib/parsers/publicsuffix-org.js
@@ -9,12 +9,12 @@ var PublicSuffixOrgParser = {};
 /**
  * Filters a commented or empty line
  *
- * @param row {String}
- * @return {String|null}
+ * @param {string} row
+ * @return {string|null}
  */
 function keepOnlyRules(row) {
   var trimmed = row.trim();
-  if (!trimmed || trimmed.indexOf('//') === 0) {
+  if (trimmed.length === 0 || trimmed.indexOf('//') === 0) {
     return null;
   }
 
@@ -27,7 +27,8 @@ function keepOnlyRules(row) {
 /**
  * Returns a rule based on string analysis
  *
- * @param rule {PublicSuffixRule}
+ * @param {string} row
+ * @return {object} a public suffix rule
  */
 function domainBuilder(row) {
   var rule = {
@@ -36,16 +37,20 @@ function domainBuilder(row) {
     parts: null,
   };
 
-  var spaceIndex = row.indexOf(' ');
   // Only read line up to the first white-space
+  var spaceIndex = row.indexOf(' ');
+  if (spaceIndex !== -1) {
+    row = row.substr(0, spaceIndex);
+  }
+
   row = punycode.toASCII(row);
 
-  //setting initial rule
+  // Keep track of initial rule
   rule.source = row;
 
-  // exceptions
+  // Exception
   if (row[0] === '!') {
-    row = row.substring(1);
+    row = row.substr(1);
     rule.exception = true;
   }