--- a/lib/public-suffix.js
+++ b/lib/public-suffix.js
@@ -1,7 +1,5 @@
 var cleanHostValue = require('./clean-host.js');
 var extractTldFromHost = require('./from-host.js');
-var getCandidateRule = require('./canditate-rule.js');
-var getRulesForTld = require('./tld-rules.js');
 
 /**
  * Returns the public suffix (including exact matches)
@@ -11,21 +9,19 @@ var getRulesForTld = require('./tld-rules.js');
  * @param {string} host
  * @return {String}
  */
-module.exports = function getPublicSuffix(allRules, host) {
-  var hostTld, rules, rule;
+module.exports = function getPublicSuffix(rules, host, dirtyHost = true) {
+  var cleanHost = dirtyHost ? cleanHostValue(host) : host;
 
-  if (host in allRules){
-	  return host;
+  // Host is a valid TLD
+  if (rules.hasTld(cleanHost)) {
+    return cleanHost;
   }
 
-  host = cleanHostValue(host);
-  hostTld = extractTldFromHost(host);
-  rules = getRulesForTld(allRules, hostTld);
-  rule = getCandidateRule(host, rules, { lazy: true });
-
-  if (rule === null) {
-    return null;
+  var candidate = rules.suffixLookup(cleanHost);
+  if (candidate === null) {
+    // Prevailing rule is '*'
+    return extractTldFromHost(cleanHost);
   }
 
-  return rule.getNormalXld().slice(1);
+  return candidate;
 };