--- a/lib/public-suffix.js
+++ b/lib/public-suffix.js
@@ -9,8 +9,8 @@ var extractTldFromHost = require('./from-host.js');
  * @param {string} host
  * @return {String}
  */
-module.exports = function getPublicSuffix(rules, host, dirtyHost = true) {
-  var cleanHost = dirtyHost ? cleanHostValue(host) : host;
+module.exports = function getPublicSuffix(rules, host, dirtyHost) {
+  var cleanHost = (dirtyHost || true) ? cleanHostValue(host) : host;
 
   // Host is a valid TLD
   if (rules.hasTld(cleanHost)) {