--- a/lib/public-suffix.js
+++ b/lib/public-suffix.js
@@ -1,3 +1,5 @@
+"use strict";
+
 var cleanHostValue = require('./clean-host.js');
 var extractTldFromHost = require('./from-host.js');
 
@@ -6,21 +8,23 @@ var extractTldFromHost = require('./from-host.js');
  *
  * @api
  * @since 1.5
- * @param {string} host
- * @return {String}
+ * @param {string} hostname
+ * @return {string}
  */
-module.exports = function getPublicSuffix(rules, host, isHostClean) {
-  var cleanHost = cleanHostValue(host, isHostClean);
+module.exports = function getPublicSuffix(rules, hostname) {
+  // Extract hostname
+  hostname = cleanHostValue(hostname);
 
-  // Host is a valid TLD
-  if (rules.hasTld(cleanHost)) {
-    return cleanHost;
+  // First check if `hostname` is already a valid top-level Domain.
+  if (rules.hasTld(hostname)) {
+    return hostname;
   }
 
-  var candidate = rules.suffixLookup(cleanHost);
+  var candidate = rules.suffixLookup(hostname);
   if (candidate === null) {
-    // Prevailing rule is '*'
-    return extractTldFromHost(cleanHost);
+    // Prevailing rule is '*' so we consider the top-level domain to be the
+    // public suffix of `hostname` (e.g.: 'example.org' => 'org').
+    return extractTldFromHost(hostname);
   }
 
   return candidate;