--- a/lib/subdomain.js
+++ b/lib/subdomain.js
@@ -1,6 +1,5 @@
 var cleanHostValue = require('./clean-host.js');
 var getDomain = require('./domain.js');
-var escapeRegExp = require('./escape-regexp.js');
 
 /**
  * Returns the subdomain of a host string
@@ -9,19 +8,14 @@ var escapeRegExp = require('./escape-regexp.js');
  * @param {string} host
  * @return {string|null} a subdomain string if any, blank string if subdomain is empty, otherwise null
  */
-module.exports = function getSubdomain(allRules, validHosts, host){
-  var domain, r, subdomain;
-
-  host = cleanHostValue(host);
-  domain = getDomain(allRules, validHosts, host);
+module.exports = function getSubdomain(allRules, validHosts, host, dirtyHost = true) {
+  var cleanHost = dirtyHost ? cleanHostValue(host) : host;
+  var domain = getDomain(allRules, validHosts, cleanHost, false);
 
   // No domain found? Just abort, abort!
-  if (domain === null){
+  if (domain === null) {
     return null;
   }
 
-  r = '\\.?'+ escapeRegExp(domain)+'$';
-  subdomain = host.replace(new RegExp(r, 'i'), '');
-
-  return subdomain;
+  return cleanHost.substring(0, cleanHost.length - domain.length - 1);
 };