--- a/lib/subdomain.js
+++ b/lib/subdomain.js
@@ -8,8 +8,8 @@ var getDomain = require('./domain.js');
  * @param {string} host
  * @return {string|null} a subdomain string if any, blank string if subdomain is empty, otherwise null
  */
-module.exports = function getSubdomain(allRules, validHosts, host, dirtyHost = true) {
-  var cleanHost = dirtyHost ? cleanHostValue(host) : host;
+module.exports = function getSubdomain(allRules, validHosts, host, dirtyHost) {
+  var cleanHost = (dirtyHost || true) ? cleanHostValue(host) : host;
   var domain = getDomain(allRules, validHosts, cleanHost, false);
 
   // No domain found? Just abort, abort!