--- a/lib/subdomain.js
+++ b/lib/subdomain.js
@@ -8,9 +8,9 @@ var getDomain = require('./domain.js');
  * @param {string} host
  * @return {string|null} a subdomain string if any, blank string if subdomain is empty, otherwise null
  */
-module.exports = function getSubdomain(allRules, validHosts, host, dirtyHost) {
-  var cleanHost = (dirtyHost || true) ? cleanHostValue(host) : host;
-  var domain = getDomain(allRules, validHosts, cleanHost, false);
+module.exports = function getSubdomain(allRules, validHosts, host, isHostClean) {
+  var cleanHost = cleanHostValue(host, isHostClean);
+  var domain = getDomain(allRules, validHosts, cleanHost, true);
 
   // No domain found? Just abort, abort!
   if (domain === null) {