--- a/lib/subdomain.js
+++ b/lib/subdomain.js
@@ -1,21 +1,26 @@
+"use strict";
+
 var cleanHostValue = require('./clean-host.js');
 var getDomain = require('./domain.js');
 
+
 /**
- * Returns the subdomain of a host string
+ * Returns the subdomain of a hostname string
  *
  * @api
- * @param {string} host
- * @return {string|null} a subdomain string if any, blank string if subdomain is empty, otherwise null
+ * @param {string} hostname
+ * @return {string|null} a subdomain string if any, blank string if subdomain
+ *  is empty, otherwise null.
  */
-module.exports = function getSubdomain(allRules, validHosts, host, isHostClean) {
-  var cleanHost = cleanHostValue(host, isHostClean);
-  var domain = getDomain(allRules, validHosts, cleanHost, true);
+module.exports = function getSubdomain(rules, validHosts, hostname) {
+  hostname = cleanHostValue(hostname);
+
+  var domain = getDomain(rules, validHosts, hostname);
 
   // No domain found? Just abort, abort!
   if (domain === null) {
     return null;
   }
 
-  return cleanHost.substring(0, cleanHost.length - domain.length - 1);
+  return hostname.substr(0, hostname.length - domain.length - 1);
 };