--- a/lib/suffix-trie.js
+++ b/lib/suffix-trie.js
@@ -17,11 +17,13 @@ function insertInTrie(rule, trie) {
 
   for (var i = 0; i < parts.length; i++) {
     var part = parts[i];
-    if (!Object.prototype.hasOwnProperty.call(node, part)) {
-      node[part] = Object.create(null);
+    var nextNode = node[part];
+    if (nextNode === undefined) {
+      nextNode = Object.create(null);
+      node[part] = nextNode;
     }
 
-    node = node[part];
+    node = nextNode;
   }
 
   node.$ = VALID_HOSTNAME_VALUE;