--- a/lib/tld-exists.js
+++ b/lib/tld-exists.js
@@ -8,8 +8,8 @@ var extractTldFromHost = require('./from-host.js');
  * @param {string} host
  * @return {boolean}
  */
-module.exports = function tldExists(rules, host, dirtyHost = true) {
-  var cleanHost = dirtyHost ? cleanHostValue(host) : host;
+module.exports = function tldExists(rules, host, dirtyHost) {
+  var cleanHost = (dirtyHost || true) ? cleanHostValue(host) : host;
   var hostTld;
 
   // Easy case, it's a TLD