… lib/canditate-rule.js
… lib/escape-regexp.js
… lib/exports/standard-json.js
… lib/rule.js
… lib/tld-rules.js
… lib/updater.js
… package.json
… test/rule.js
… test/tld.js