Hi there,

I spent some time this week-end implementing a different representation for the rules: replacing the regex by a more efficient Trie data structure. This changes a lot of code, so I would understand that it might require more work to make "production-ready". Also, comments are pretty sparse at the moment, but I would be glad to provide more where you think it makes sense!

I implemented some benchmark to have an idea of the performance gain, and it looks promising (although I did not register the memory consumption, I don't expect it to be unreasonable). Each `run` is calling the benchmarked function on 100 random domains and I'm using the `benchmark` npm package (I can attach the code somewhere as well if you want to have a look):

## Current version of the code on master (v2.0.0)

```
tldjs#tldExists x 3,162 ops/sec ±1.74% (92 runs sampled)
tldjs#getDomain x 35.82 ops/sec ±1.65% (62 runs sampled)
tldjs#getSubdomain x 35.77 ops/sec ±1.60% (62 runs sampled)
tldjs#getPublicSuffix x 37.39 ops/sec ±1.13% (64 runs sampled)
```

## New implementation based on the Trie data-structure

```
tldjs#tldExists x 3,265 ops/sec ±0.83% (92 runs sampled)
tldjs#getDomain x 2,030 ops/sec ±3.08% (84 runs sampled)
tldjs#getSubdomain x 2,037 ops/sec ±2.85% (85 runs sampled)
tldjs#getPublicSuffix x 2,195 ops/sec ±3.53% (83 runs sampled)
```

So the speed-up is around `x60`.

It now seems that `cleanHostValue` is the bottle-neck, since if we make it a no-op, the results become:

```
tldjs#tldExists x 57,589 ops/sec ±2.23% (92 runs sampled)
tldjs#getDomain x 8,414 ops/sec ±0.75% (94 runs sampled)
tldjs#getSubdomain x 8,257 ops/sec ±0.77% (94 runs sampled)
tldjs#getPublicSuffix x 10,857 ops/sec ±1.27% (92 runs sampled)
```

I added a flag (`dirtyHost`) on each of these methods to allow disabling the cleaning (if you know that your host value is already clean and you want maximum performances).

While implementing this new representation, I stumbled upon some cases for which I'm not really sure of the good behavior. I would love to have your feedback on them.

## Result if the public suffix does not exist

The following test-case: `expect(tld.getPublicSuffix('www.freedom.nsa')).to.be(null);` seems to expect the wrong result. According to the specification (https://publicsuffix.org/list/). When no matching rule is found, then "the prevailing rule is "*".", which from what I understand means that we consider the latest label of the hostname to be the public suffix. The result for this test would then be `nsa` instead of `null`. What do you think?

## Result of exceptions

The following test-case: `expect(tld.getPublicSuffix('www.www.ck')).to.be('www.ck');` should probably expect `ck` instead of `www.ck` as a public suffix. According to the aforementioned specification, the following should be done when an exception is matched (here: `!www.ck`):
> If the prevailing rule is a exception rule, modify it by removing the leftmost label.

I guess we should then remove `www` (the leftmost label), resulting in a public suffix of `ck`.

## Result when not public suffix is found

What should be the result of `getDomain` when no valid public suffix is found? Should it be `null` as well, or the value passed as `host` to the function?

## Ignoring trailing and leading dots

The specification says that trailing and leading dots found on rules should be ignored. Although there is no such case at the moment, maybe we should handle them and add a test-case? I left a `TODO` in the parser.

Thanks for reading, and I hope this work can be useful to the project,
Have a nice evening.