@oncletom 2017-07-24
Hey @remusao,

thanks for the hard work! Figures are impressive 🙂
I will try to have a close look at it tomorrow—I am viewing houses in London at the moment and some transcripts take the rest of my time at the moment.

> When no matching rule is found, then "the prevailing rule is "*".", which from what I understand means that we consider the latest label of the hostname to be the public suffix. The result for this test would then be `nsa` instead of `null`. What do you think?

Yeah, I feel it would be nicer to respect what the public suffix spec says — I was feeling frustrated about the methods, what they were doing and for what in #25 (#54 maybe). It started from "using public-suffix.org to enforce a verification of known TLDs" to "implementing PublicSuffix for cookies and stuff while trying to address the former intention".