@remusao 2017-07-24
Hey @oncletom,

Thanks for the answer. Let me know if I can do anything to ease the review (more comments, a tl;dr about the implementation details, or more direct talk are possible!).

Regarding the issues you mention, I was thinking about using only the ICANN section of the file to get the domain (and use the PRIVATE section only to get the public suffix). Do you think it could work?