@oncletom 2017-08-03
@remusao ICANN is a good idea! Or IANA (cf. #40) too – I don't know which one validates gTLD first. It would be useful if the Public Suffix list does not encompass all the gTLDs or with a major delay.

I found a place to hopefully Sunday I should be able to put my head around your proposal. Will let you know if I have any question or else — if other contributor(s) want to have a peak at it, they are obviously welcome to do so 🙂