@remusao 2017-08-08
@oncletom Thanks for your reply! I understand that time is precious and this is a big PR. If I can do anything to make it easier to review, let me know.