@ctavan 2017-08-09
@remusao since I'm very interested in this PR getting merged I tried to verify the two cases that you highlight in the issue description.

## www.freedom.nsa

I think your reasoning is right, the public suffix should be `nsa`. To verify the reasoning I simply took the definitions from https://publicsuffix.org/list/ and highlighted the essential parts:

### Definitions

- [ … ]
- A domain is said to match a rule if and only if all of the following conditions are met:
  - When the domain and rule are split into corresponding labels, that the domain contains as many or more labels than the rule.
  - **Beginning with the right-most labels of both the domain and the rule, and continuing for all labels in the rule, one finds that for every pair, either they are identical, or that the label from the rule is "*"**.

### Algorithm

1. Match domain against all rules and take note of the matching ones.
2. **If no rules match, the prevailing rule is "*".**
3. If more than one rule matches, the prevailing rule is the one which is an exception rule.
4. If there is no matching exception rule, the prevailing rule is the one with the most labels.
5. If the prevailing rule is a exception rule, modify it by removing the leftmost label.
6. **The public suffix is the set of labels from the domain which match the labels of the prevailing rule**, using the matching algorithm above.
7. The registered or registrable domain is the public suffix plus one additional label.

```
prevailing rule = *

Domain: | www | freedom | nsa |
Rule:   |     |         |  *  |

public suffix = nsa
```

## www.www.ck

I think your reasoning is again right (public suffix should be `ck`):

1. Match domain against all rules and take note of the matching ones.
   - Matching rules:

         *.ck
         !www.ck

2. ~~If no rules match, the prevailing rule is "*".~~
3. If more than one rule matches, the prevailing rule is the one which is an exception rule.
   - Matching rule:

         !www.ck

4. ~~If there is no matching exception rule, the prevailing rule is the one with the most labels.~~
5. If the prevailing rule is a exception rule, modify it by removing the leftmost label.
   - Modified matching rule:

         ck

6. The public suffix is the set of labels from the domain which match the labels of the prevailing rule, using the matching algorithm above.
7. The registered or registrable domain is the public suffix plus one additional label.

```
prevailing rule = ck

Domain: | www | www | ck |
Rule:   |     |     | ck |

public suffix = ck
registered domain = www.ck
```