@oncletom 2017-08-31
@remusao I have been thinking about the _clean host_ vs. _dirty host_. I agree there is a need for cleaning if it is relevant only. But by changing the interface we can no longer trust the flag which states if the value is clean or not.

In my opinion, the `cleanHost` function should return a frozen `CleanHostString` object. Maybe like this:

```js
class CleanHostString extends String {
  constructor(value) {
    super(value);

    this.hostValue = value instanceof CleanHost ? String(value) : cleanHost(value);

    // we ensure the object becomes immutable
    Object.freeze(this);
  }

  // returns the value when the conversion to String primitive happens, with `String(obj)` 
  valueOf() {
    return this.hostValue;
  }
}
```

What do you think?