@remusao 2017-08-31
@oncletom Yeah the flag might not be the ideal solution. What about just asking users to clean the url before giving it to functions of tldjs? So that we simplify the API and all functions expect an already clean host.

Another option could be to wrap clean host in a different type and check from each function if we are given an already clean host, or if the cleaning needs to happen? But then it's not propagated to the caller and we still might have redundancy. Although I guess it's nice because we hide all the complexity from the user and it's more efficient than what we have now (but then users cannot chose to not have any pre-processing at all, or do it themselves, is it something we want to offer?).

Also, on a side note, I started working on a more efficient version of `cleanHost` to limit its impact on the overall performance of the library.

Thanks for your answer! If you think that the best option is to wrap the clean host in a different type, I can implement it in this PR.

Also, I see you make use of `class` in your example, and it made me wonder what version of Javascript this library should support? `es5`, `es6`, something else?