@remusao 2017-08-31
@oncletom No problem for the delay, thank you so much for reviewing. I will address your comment and add some changes.

One thing that remains is what to do regarding the two failing tests, should I change them to adopt a behavior as described by the specification?

> I have not thought much about it — I realised only after publishing the comment the codebase was in ES5 😄 If the trie implementation does not break anything in the API, I'd rather keep it as is and later do a major bump to move to ES2015, when Node 8 is out for example.

Makes sense, I will keep the code compatible (the trie should not break anything).