@oncletom 2017-08-31
> One thing that remains is what to do regarding the two failing tests, should I change them to adopt a behavior as described by the specification?

Yes, please do proceed as is 🙂 it is totally fine to fix bugs along the way!